Add files via upload
Main Goal:-

The VPC creates a private network for your instances. Although the Application Load Balancer (ALB) is deployed in a public subnet, the instances are placed in private subnets, making them inaccessible directly from the internet. This setup ensures that only the ALB can communicate with the instances.
For example, think of the VPC as a gated community—while the ALB acts as the front gate accessible to the public, the instances (your application or website) reside securely inside. When an external user accesses the application through the ALB’s public DNS, the ALB forwards the request to the instances in the private subnet and returns the response.
Thus, users do not directly access the instances; they interact with the ALB, which securely handles communication with the backend instances within the same VPC.